chore: move unit tests into a unit folder
This allows packages to more easily separate unit and integration tests.
